+++ src/main/kotlin/com/jetbrains/snakecharm/inspections/SmkTemplateEngineSettingsInspection.kt
@@ -4,11 +4,11 @@ import com.intellij.codeInspection.LocalInspectionToolSession
 import com.intellij.codeInspection.ProblemsHolder
 import com.jetbrains.python.psi.PyStringLiteralExpression
 import com.jetbrains.snakecharm.SnakemakeBundle
-import com.jetbrains.snakecharm.codeInsight.completion.ShadowSectionSettingsProvider
+import com.jetbrains.snakecharm.codeInsight.completion.TemplateEngineSettingsProvider
 import com.jetbrains.snakecharm.lang.SnakemakeNames
 import com.jetbrains.snakecharm.lang.psi.SmkRuleOrCheckpointArgsSection
 
-class SmkShadowSettingsInspection : SnakemakeInspection() {
+class SmkTemplateEngineSettingsInspection : SnakemakeInspection() {
     override fun buildVisitor(
         holder: ProblemsHolder,
         isOnTheFly: Boolean,
@@ -16,19 +16,21 @@ class SmkShadowSettingsInspection : SnakemakeInspection() {
     ) = object : SnakemakeInspectionVisitor(holder, getContext(session)) {
 
         override fun visitSmkRuleOrCheckpointArgsSection(st: SmkRuleOrCheckpointArgsSection) {
-            if (st.name != SnakemakeNames.SECTION_SHADOW) {
+            if (st.name != SnakemakeNames.SECTION_TEMPLATE_ENGINE) {
                 return
             }
 
             val argument = st.argumentList?.arguments?.firstOrNull()
             if (argument is PyStringLiteralExpression &&
-                argument.stringValue !in ShadowSectionSettingsProvider.OPTIONS
+                argument.stringValue !in TemplateEngineSettingsProvider.OPTIONS
             ) {
-                registerProblem(argument.originalElement,
-                    SnakemakeBundle.message("INSP.NAME.shadow.settings.msg"))
+                registerProblem(
+                    argument.originalElement,
+                    SnakemakeBundle.message("INSP.NAME.template.engine.settings.msg")
+                )
             }
         }
     }
 
-    override fun getDisplayName(): String = SnakemakeBundle.message("INSP.NAME.shadow.settings.title")
+    override fun getDisplayName(): String = SnakemakeBundle.message("INSP.NAME.template.engine.settings.title")
 }