+++ src/main/kotlin/com/jetbrains/snakecharm/lang/psi/impl/SmkRuleOrCheckpointArgsSectionImpl.kt
@@ -49,6 +49,7 @@ open class SmkRuleOrCheckpointArgsSectionImpl(node: ASTNode) : SmkArgsSectionImp
                 )
             }
 
+            // TODO: SECTION_TEMPLATE_ENGINE
             else -> null
         }